Added shell tests and reorganized libraries
