--- app/src/sensor/bme688.h
@@ -1,7 +1,0 @@
-
-#ifndef __BME688_ENTRY_H__
-#define __BME688_ENTRY_H__
-
-void bme_entry_point(void *_msgq, void *arg2, void *arg3);
-
-#endif /* __BME688_ENTRY_H__ */
