+++ app/src/sensor/sps30.c
@@ -4,8 +4,7 @@
 #include <logging/log.h>
 #include <zephyr.h>
 #include <stdlib.h>
-#include "bme688.h"
-#include "cayenne.h"
+#include <cayenne.h>
 
 LOG_MODULE_DECLARE(aether);
