--- app/src/sensor/sps30.h
@@ -1,7 +1,0 @@
-
-#ifndef __SPS30_ENTRY_H__
-#define __SPS30_ENTRY_H__
-
-void sps_entry_point(void *_msgq, void *arg2, void *arg3);
-
-#endif /* __SPS30_ENTRY_H__ */
