+++ app/src/sensor/zmod4510.c
@@ -5,8 +5,7 @@
 #include <logging/log.h>
 #include <zephyr.h>
 #include <stdlib.h>
-#include "bme688.h"
-#include "cayenne.h"
+#include <cayenne.h>
 
 LOG_MODULE_DECLARE(aether);
