--- app/src/sensor/zmod4510.h
@@ -1,7 +1,0 @@
-
-#ifndef __ZMOD4510_ENTRY_H__
-#define __ZMOD4510_ENTRY_H__
-
-void zmod_entry_point(void *_msgq, void *arg2, void *arg3);
-
-#endif /* __ZMOD4510_ENTRY_H__ */
