--- app/src/threads.h
@@ -1,5 +1,0 @@
-/*************************** Thread Definitions *******************************/
-
-#define NORMAL_PRIORITY		5
-#define HIGH_PRIORITY		3
-
