+++ lib/cayenne/CMakeLists.txt
@@ -1,0 +1,2 @@
+
+zephyr_sources(cayenne.c)
