+++ tests/lib/cayenne/Kconfig
@@ -1,0 +1,1 @@
+CONFIG_ZTEST=y
